pyproject.toml
[tool.poetry]
name = "micro-flask"
version = "0.1.0"
description = ""
authors = ["LeSavantR <ruben_hernandez1993@hotmail.es>"]
readme = "README.md"
packages = [{include = "micro_flask"}]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.9.4"
autoenv = "^1.0.0"
blinker = "^1.5"
certifi = "^2022.12.7"
cffi = "^1.15.1"
chardet = "^5.1.0"
click = "^8.1.3"
cryptography = "^39.0.1"
dnspython = "^2.3.0"
dominate = "^2.7.0"
flask = "^2.2.3"
flask-bootstrap = "^3.3.7.1"
flask-dotenv = "^0.1.2"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
flask-sqlalchemy = "^3.0.3"
flask-uploads = "^0.2.1"
flask-wtf = "^1.1.1"
idna = "^3.4"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
mako = "^1.2.4"
markupsafe = "^2.1.2"
marshmallow = "^3.19.0"
passlib = "^1.7.4"
protobuf = "^4.22.0"
pycparser = "^2.21"
pymysql = "^1.0.2"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
python-editor = "^1.0.4"
requests = "^2.28.2"
six = "^1.16.0"
sqlalchemy = "^2.0.4"
urllib3 = "^1.26.14"
visitor = "^0.1.3"
werkzeug = "^2.2.3"
wtforms = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
